core: create suiteSubprocess module
Move the setup code for the suite subprocess into another module to
facilitate importing and using the Suite class directly

Change-type: patch
