+++ core/lib/main.js
@@ -245,7 +245,7 @@ async function setup() {
 
 				// The reason we need to fork is because many 3rd party libariers output to stdout
 				// so we need to capture that
-				suite = fork('./lib/common/suite', {
+				suite = fork('./lib/common/suiteSubprocess', {
 					stdio: ['pipe', 'pipe', 'pipe', 'ipc'],
 				});
 				suiteStarted = true;
